backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
numpy==1.26.4
opencv-python-headless==4.8.1.76
scikit-image==1.1.0
Pillow==10.1.0
python-multipart==0.0.6
